start backend migrations from scratch
There's a chicken/egg situation where a table is changed to point to Source.id
but Source's id is renamed from source_id later on, and Alembic haven't
figured out the ordering to make this work.
I'd rather also not modify migrations by hand, is this is likely not the
only instance of this problem.