Add support for s3:
This code is admittedly a little hacky/non-pythonic:
- a better wrapper could be written instead of multiple conditionals.
- instead of relying on click to create a path type for us, we explicitly create it
because sometimes we want to use a cloud path. This could likely be dealt with more
neatly with a custom type, but likely overkill for now.
- if s3 is used, we create a temporary directory to write memmap files to. These are deleted
at the end after we've sent the embeddings to an s3 object.
- There's now an s3 option. This can be inferred from the file path, but this makes it more explicit.
Strictly speaking this option could depend on other options, but again overkill for now.