Add facilities to merge in style blocks that Adobe cannot detect.
- Since Adobe fails to detect some styling elements (italics, but also sometimes other elements), these elements have different paths
and so are not included by simple "merge duplicated path" logic.

- This commit solves this issue by assuming that blocks between text blocks with the same path are also part of those semantics.

- Unfortunately, the facility doesn't allow us to identify the style types in these cases. Instead, I assign the last element of the path as the
style type (usually Span). This doesn't usually matter for semantics, however, so we shouldn't worry. This metadata is helpful for debugging (it
tells us that Adobe gets things wrong alarmingly often, something like 20% of passages with style aren't captured!)